# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegistrationInput {
  /** @minLength 2 */
  fullName: string;
  /** @minLength 3 */
  username: string;
  email: string;
  confirmEmail: string;
  phoneCountryCode: string;
  phoneNumber: string;
  /** @nullable */
  referralCode?: string | null;
  /** @minLength 6 */
  password: string;
  /** @minLength 6 */
  confirmPassword: string;
}

export interface RegistrationResponse {
  message: string;
  email: string;
}

export interface VerifyInput {
  email: string;
  /**
     * @minLength 6
     * @maxLength 6
     */
  code: string;
}

export interface UserProfile {
  id: number;
  fullName: string;
  username: string;
  email: string;
  phoneCountryCode?: string;
  phoneNumber?: string;
  /** @nullable */
  referralCode?: string | null;
  createdAt?: string;
}

export interface VerifyResponse {
  message: string;
  user: UserProfile;
}

export interface ResendInput {
  email: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface LoginResponse {
  message: string;
  user: UserProfile;
}

export interface ForgotPasswordInput {
  email: string;
}

export interface ResetPasswordInput {
  email: string;
  /**
     * @minLength 6
     * @maxLength 6
     */
  code: string;
  /** @minLength 6 */
  newPassword: string;
  /** @minLength 6 */
  confirmNewPassword: string;
}

export interface MessageResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
}