# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication and registration
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user and send confirmation code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationInput"
      responses:
        "200":
          description: Confirmation code sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email or username already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify:
    post:
      operationId: verifyCode
      tags: [auth]
      summary: Verify the 6-digit confirmation code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyInput"
      responses:
        "200":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyResponse"
        "400":
          description: Invalid or expired code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/resend:
    post:
      operationId: resendCode
      tags: [auth]
      summary: Resend confirmation code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResendInput"
      responses:
        "200":
          description: Code resent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationResponse"
        "400":
          description: No pending registration for this email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request a password reset code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Reset code sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password using the code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid or expired code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]
    RegistrationInput:
      type: object
      required:
        - fullName
        - username
        - email
        - confirmEmail
        - phoneCountryCode
        - phoneNumber
        - password
        - confirmPassword
      properties:
        fullName:
          type: string
          minLength: 2
        username:
          type: string
          minLength: 3
        email:
          type: string
          format: email
        confirmEmail:
          type: string
          format: email
        phoneCountryCode:
          type: string
        phoneNumber:
          type: string
        referralCode:
          type: ["string", "null"]
        password:
          type: string
          minLength: 6
        confirmPassword:
          type: string
          minLength: 6
    RegistrationResponse:
      type: object
      required: [message, email]
      properties:
        message:
          type: string
        email:
          type: string
    VerifyInput:
      type: object
      required: [email, code]
      properties:
        email:
          type: string
          format: email
        code:
          type: string
          minLength: 6
          maxLength: 6
    VerifyResponse:
      type: object
      required: [message, user]
      properties:
        message:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
    ResendInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    LoginResponse:
      type: object
      required: [message, user]
      properties:
        message:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
    ForgotPasswordInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ResetPasswordInput:
      type: object
      required: [email, code, newPassword, confirmNewPassword]
      properties:
        email:
          type: string
          format: email
        code:
          type: string
          minLength: 6
          maxLength: 6
        newPassword:
          type: string
          minLength: 6
        confirmNewPassword:
          type: string
          minLength: 6
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    UserProfile:
      type: object
      required: [id, fullName, username, email]
      properties:
        id:
          type: integer
        fullName:
          type: string
        username:
          type: string
        email:
          type: string
        phoneCountryCode:
          type: string
        phoneNumber:
          type: string
        referralCode:
          type: ["string", "null"]
        createdAt:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string