# lib/api-zod/src/generated/types/resetPasswordInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ResetPasswordInput {
  email: string;
  /**
     * @minLength 6
     * @maxLength 6
     */
  code: string;
  /** @minLength 6 */
  newPassword: string;
  /** @minLength 6 */
  confirmNewPassword: string;
}